# packages/contract/openapi.yaml
openapi: "3.1.0"
info:
  title: Statics Platform API
  version: "0.1.0"
  description: >
    API contract for the Statics Problem Platform.
    This file is the source of truth — edit it before implementing either side.
    Regenerate TS types with: pnpm --filter @statics/contract generate
servers:
  - url: http://localhost:8000
    description: Local development
paths:
...
  /api/v1/assignments/{slug}:
    get:
      operationId: getAssignment
      summary: Get assignment summary
      tags: [assignments]
      parameters:
        - $ref: "#/components/parameters/slug"
      responses:
        "200":
          description: Assignment summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssignmentSummary"
        "404":
          description: Assignment not found
...
  /api/v1/admin/problem-types:
    get:
      operationId: listProblemTypes
      summary: Registered problem types and their configurable params
      tags: [admin-assignments]
      responses:
        "200":
          description: Problem type catalogue
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProblemTypeInfo"
        "401":
          description: Not authenticated
  /api/v1/admin/courses/{courseId}/assignments:
    get:
      operationId: listCourseAssignments
      summary: Assignments in a course
      tags: [admin-assignments]
      parameters:
        - $ref: "#/components/parameters/courseId"
      responses:
        "200":
          description: Assignment list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminAssignmentSummary"
        "404":
          description: No such course
    post:
      operationId: createAssignment
      summary: Create a draft assignment with its problem slots
      tags: [admin-assignments]
      parameters:
        - $ref: "#/components/parameters/courseId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignmentCreateRequest"
      responses:
        "201":
          description: Draft created — not visible to students until published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAssignmentDetail"
        "403":
          description: Insufficient course role
        "409":
          description: That slug is already used in this course
  /api/v1/admin/assignments/{assignmentId}:
    get:
      operationId: getAdminAssignment
      summary: Assignment detail, including problem slots and targeting
      tags: [admin-assignments]
      parameters:
        - $ref: "#/components/parameters/assignmentId"
      responses:
        "200":
          description: Assignment detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAssignmentDetail"
        "404":
          description: No such assignment, or caller has no role on its course
    patch:
      operationId: updateAssignment
      summary: Update settings, problem slots and/or targeting
      description: >
        Every field is optional; omitted fields are left alone. `problems` and
        `targets` replace their collection wholesale when present.
      tags: [admin-assignments]
      parameters:
        - $ref: "#/components/parameters/assignmentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignmentUpdateRequest"
      responses:
        "200":
          description: Updated assignment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAssignmentDetail"
        "403":
          description: Insufficient course role
        "404":
          description: No such assignment
        "409":
          description: That slug is already used in this course
  /api/v1/admin/assignments/{assignmentId}/publish:
    post:
      operationId: publishAssignment
      summary: Publish or unpublish an assignment
      description: >
        Publishing is what makes an assignment visible to targeted students.
        Rejected when the assignment has no problem slots, since a published
        empty assignment is a dead end for the student.
      tags: [admin-assignments]
      parameters:
        - $ref: "#/components/parameters/assignmentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishRequest"
      responses:
        "200":
          description: Updated assignment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAssignmentDetail"
        "403":
          description: Insufficient course role
        "409":
          description: Cannot publish an assignment with no problems
        "404":
          description: No such assignment
  /api/v1/admin/assignments/{assignmentId}/preview/{index}:
    get:
      operationId: previewAssignmentProblem
      summary: Render one problem slot as a student would see it
      description: >
        Instructor-only. Returns the same payload shape a student receives, so
        the builder can reuse the problem renderer. Carries no solution data —
        the separate replay route is the only solution-bearing endpoint.
      tags: [admin-assignments]
      parameters:
        - $ref: "#/components/parameters/assignmentId"
        - $ref: "#/components/parameters/index"
        - name: seed
          in: query
          required: false
          schema:
            type: integer
          description: Reroll the preview with a specific seed
      responses:
        "200":
          description: Problem payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemPayload"
        "404":
          description: No such assignment or problem index
...
components:
  parameters:
    slug:
      name: slug
      in: path
      required: true
      schema:
        type: string
      description: Assignment slug (e.g. "truss-fall-2026")
    index:
      name: index
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
      description: 1-based problem index within the assignment
    courseId:
      name: courseId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Internal course UUID
    assignmentId:
      name: assignmentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Internal assignment UUID
...
  schemas:
...
    ParamFieldSpec:
      type: object
      description: One configurable generator knob, described so a UI can render it
      required: [name, label, valueType, default, minimum, maximum, step, helpText]
      properties:
        name:
          type: string
          description: Key written into the problem slot's params
        label:
          type: string
        valueType:
          type: string
          enum: [integer, number]
        default:
          type: number
          description: Used when the instructor leaves the field alone
        minimum:
          type: number
          nullable: true
        maximum:
          type: number
          nullable: true
        step:
          type: number
          nullable: true
        helpText:
          type: string
    ProblemTypeInfo:
      type: object
      required: [problemType, displayName, paramsSchema]
      properties:
        problemType:
          type: string
          description: Drives the renderer registry
          example: truss
        displayName:
          type: string
          example: Planar truss
        paramsSchema:
          type: array
          items:
            $ref: "#/components/schemas/ParamFieldSpec"
    FeedbackMode:
      type: string
      enum: [binary, per_field]
    AssignmentAudience:
      type: string
      description: >
        `all` is evaluated at request time, so students who join the course
        later are included without re-assigning. `selected` uses an explicit
        target list.
      enum: [all, selected]
    AdminProblemSlot:
      type: object
      required: [orderIndex, problemType, params, points]
      properties:
        orderIndex:
          type: integer
          minimum: 1
          description: 1-based position in the assignment
        problemType:
          type: string
        params:
          type: object
          additionalProperties:
            type: number
          description: Keys come from the problem type's paramsSchema
        points:
          type: number
          default: 1
    AdminAssignmentSummary:
      type: object
      required:
        [id, courseId, slug, title, isPublished, audience, problemCount,
         targetedStudentCount, dueAt, opensAt, hardDeadlineAt, allowLate,
         latePenaltyRate, revealSolutionsAfterClose]
      properties:
        id:
          type: string
          format: uuid
        courseId:
          type: string
          format: uuid
        slug:
          type: string
        title:
          type: string
        isPublished:
          type: boolean
        audience:
          $ref: "#/components/schemas/AssignmentAudience"
        problemCount:
          type: integer
        targetedStudentCount:
          type: integer
        dueAt:
          type: string
          format: date-time
          nullable: true
        opensAt:
          type: string
          format: date-time
          nullable: true
        hardDeadlineAt:
          type: string
          format: date-time
          nullable: true
          description: Absolute cutoff; only meaningful when allowLate is true
        allowLate:
          type: boolean
        latePenaltyRate:
          type: number
        revealSolutionsAfterClose:
          type: boolean
          description: >
            When true, students may see reference solutions once the assignment
            has closed. Default false — solutions never leave the backend.
    AdminAssignmentDetail:
      allOf:
        - $ref: "#/components/schemas/AdminAssignmentSummary"
        - type: object
          required:
            [instructions, tolerance, feedbackMode, maxAttempts, problems, targetEntryIds]
          properties:
            instructions:
              type: string
            tolerance:
              type: number
            feedbackMode:
              $ref: "#/components/schemas/FeedbackMode"
            maxAttempts:
              type: integer
              nullable: true
              description: null means unlimited
            problems:
              type: array
              items:
                $ref: "#/components/schemas/AdminProblemSlot"
            targetEntryIds:
              type: array
              description: Roster-entry ids; only meaningful when audience is `selected`
              items:
                type: string
                format: uuid
...
    AssignmentCreateRequest:
      type: object
      required: [title, slug]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        slug:
          type: string
          minLength: 1
          maxLength: 128
        instructions:
          type: string
        dueAt:
          type: string
          format: date-time
          nullable: true
        opensAt:
          type: string
          format: date-time
          nullable: true
        hardDeadlineAt:
          type: string
          format: date-time
          nullable: true
          description: Absolute cutoff; only meaningful when allowLate is true
        allowLate:
          type: boolean
        latePenaltyRate:
          type: number
        revealSolutionsAfterClose:
          type: boolean
          description: >
            When true, students may see reference solutions once the assignment
            has closed. Default false — solutions never leave the backend.
        tolerance:
          type: number
        feedbackMode:
          $ref: "#/components/schemas/FeedbackMode"
        maxAttempts:
          type: integer
          nullable: true
        audience:
          $ref: "#/components/schemas/AssignmentAudience"
        targetEntryIds:
          type: array
          items:
            type: string
            format: uuid
        problems:
          type: array
          items:
            $ref: "#/components/schemas/AdminProblemSlot"
    AssignmentUpdateRequest:
      type: object
      description: All fields optional; omitted fields are left unchanged
      properties:
        title:
          type: string
        slug:
          type: string
        instructions:
          type: string
        dueAt:
          type: string
          format: date-time
          nullable: true
        opensAt:
          type: string
          format: date-time
          nullable: true
        hardDeadlineAt:
          type: string
          format: date-time
          nullable: true
          description: Absolute cutoff; only meaningful when allowLate is true
        allowLate:
          type: boolean
        latePenaltyRate:
          type: number
        revealSolutionsAfterClose:
          type: boolean
          description: >
            When true, students may see reference solutions once the assignment
            has closed. Default false — solutions never leave the backend.
        tolerance:
          type: number
        feedbackMode:
          $ref: "#/components/schemas/FeedbackMode"
        maxAttempts:
          type: integer
          nullable: true
        audience:
          $ref: "#/components/schemas/AssignmentAudience"
        targetEntryIds:
          type: array
          items:
            type: string
            format: uuid
        problems:
          type: array
          items:
            $ref: "#/components/schemas/AdminProblemSlot"
    PublishRequest:
      type: object
      required: [isPublished]
      properties:
        isPublished:
          type: boolean
...
    AssignmentSummary:
      type: object
      required: [slug, title, problemCount, status, feedbackMode, maxAttempts, dueAt, locked, problems, lockReason, closesAt, revealSolutions]
      properties:
        slug:
          type: string
        title:
          type: string
        problemCount:
          type: integer
        status:
          type: string
          enum: [not_started, in_progress, submitted, closed]
          description: >
            `closed` means the deadline passed without an explicit submit; the
            work is final and read-only but was never hand-submitted.
        feedbackMode:
          type: string
          enum: [binary, per_field]
        maxAttempts:
          type: integer
          nullable: true
        dueAt:
          type: string
          format: date-time
          nullable: true
        locked:
          type: boolean
          description: True once submitted or past the effective close time
        lockReason:
          type: string
          enum: [submitted, past_due]
          nullable: true
          description: Why the assignment is read-only; null while open
        closesAt:
          type: string
          format: date-time
          nullable: true
          description: >
            Effective close time under the late policy. null means the
            assignment never auto-closes.
        revealSolutions:
          type: boolean
          description: >
            True only when the assignment is closed AND the instructor enabled
            solution reveal. Never a rule the client enforces — it only tells
            the UI whether `correctAnswers` will be populated.
        problems:
          type: array
          items:
            $ref: "#/components/schemas/ProblemSummary"
        score:
          type: object
          nullable: true
          required: [earned, total]
          properties:
            earned:
              type: number
            total:
              type: number
    ProblemSummary:
      type: object
      required: [index, status, attemptCount, locked]
      properties:
        index:
          type: integer
        status:
          type: string
          enum: [no_attempt, incorrect, correct]
        attemptCount:
          type: integer
        locked:
          type: boolean
          description: True when already answered correctly or the assignment closed

    # -------------------------------------------------------------------------
    # Problem payload
    # -------------------------------------------------------------------------
    ProblemPayload:
      type: object
      required: [schemaVersion, index, problemType, prompt, geometry, answerSchema, savedAnswers, attemptCount, maxAttempts, locked, lockReason, correctAnswers]
      properties:
        schemaVersion:
          type: integer
          enum: [1]
        index:
          type: integer
        problemType:
          type: string
          description: Drives the renderer registry (e.g. "truss")
        prompt:
          $ref: "#/components/schemas/Prompt"
        geometry:
          description: Opaque to the shell — typed per problemType by the renderer
        answerSchema:
          $ref: "#/components/schemas/AnswerSchema"
        savedAnswers:
          type: object
          additionalProperties:
            type: number
            nullable: true
        attemptCount:
          type: integer
        maxAttempts:
          type: integer
          nullable: true
        locked:
          type: boolean
          description: >
            True when this problem is read-only — either already answered
            correctly, or the whole assignment closed.
        lockReason:
          type: string
          enum: [correct, submitted, past_due, unavailable]
          nullable: true
          description: >
            `unavailable` means this problem type has no server-side solver
            yet, so it can be displayed but not answered.
        correctAnswers:
          type: object
          nullable: true
          additionalProperties:
            type: number
          description: >
            Reference solution, keyed like `savedAnswers`. Populated ONLY when
            the assignment is closed AND the instructor enabled solution
            reveal; null in every other case.
    Prompt:
      type: object
      required: [title, body, notes]
      properties:
        title:
          type: string
        body:
          type: string
        notes:
          type: array
          items:
            type: string
    AnswerSchema:
      type: object
      required: [groups]
      properties:
        groups:
          type: array
          items:
            $ref: "#/components/schemas/AnswerGroup"
    AnswerGroup:
      type: object
      required: [id, label, fields]
      properties:
        id:
          type: string
        label:
          type: string
        fields:
          type: array
          items:
            $ref: "#/components/schemas/AnswerField"
    AnswerField:
      type: object
      required: [key, label, type, decimals]
      properties:
        key:
          type: string
          description: Matches keys in savedAnswers and CheckResult.perField
        label:
          type: string
          description: Rendered label (e.g. "S₁")
        unit:
          type: string
          nullable: true
          example: F
        type:
          type: string
          enum: [number]
        decimals:
          type: integer
          nullable: true
          example: 2

    # -------------------------------------------------------------------------
    # TrussGeometry — concrete geometry type for problemType="truss"
    # (geometry field in ProblemPayload is opaque; this is the resolved shape)
    # -------------------------------------------------------------------------
...